Move test classes to the right package
Test classes are now in the same package than the tested class.

Fake activities are now in the tests.common bundle, which will
allow to share them with future test fragments.
