--- tests/fr.kazejiyu.ekumi.core.test/src/fr/kazejiyu/ekumi/test/mocks/WaitFor.java
@@ -1,26 +1,0 @@
-package fr.kazejiyu.ekumi.test.mocks;
-
-import fr.kazejiyu.ekumi.core.ekumi.Context;
-import fr.kazejiyu.ekumi.core.ekumi.impl.ActivityImpl;
-
-public class WaitFor extends ActivityImpl {
-	
-	private final long delay;
-	
-	public WaitFor(long delay) {
-		super();
-		this.delay = delay;
-	}
-	
-	
-	@Override
-	public void run(Context context) {
-		try {
-			Thread.sleep(delay);
-		} catch (InterruptedException e) {
-			Thread.currentThread().interrupt();
-			throw new RuntimeException("Thread interrupted while sleeping", e);
-		}
-	}
-
-}
