--- tests/fr.kazejiyu.ekumi.core.test/src/fr/kazejiyu/ekumi/test/mocks/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Defines mocks easing tests definition.
- * 
- * @author Emmanuel CHEBBI 
- */
-package fr.kazejiyu.ekumi.test.mocks;
