--- tests/fr.kazejiyu.ekumi.core.test/src/fr/kazejiyu/ekumi/test/unit/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Defines unit tests verifying EKumi classes behavior.
- * 
- * @author Emmanuel CHEBBI
- */
-package fr.kazejiyu.ekumi.test.unit;
