+++ tests/fr.kazejiyu.ekumi.tests.common/META-INF/MANIFEST.MF
@@ -18,5 +18,7 @@ Require-Bundle: org.junit.jupiter.api;bundle-version="5.0.0";visibility:=reexpor
  org.objenesis;bundle-version="1.0.0";visibility:=reexport,
  org.hamcrest.core;bundle-version="1.3.0";visibility:=reexport,
  org.mockito;bundle-version="2.13.0";visibility:=reexport,
- net.bytebuddy.byte-buddy;bundle-version="1.7.9";visibility:=reexport
-Export-Package: fr.kazejiyu.ekumi.tests.common.mock
+ net.bytebuddy.byte-buddy;bundle-version="1.7.9";visibility:=reexport,
+ fr.kazejiyu.ekumi.core
+Export-Package: fr.kazejiyu.ekumi.tests.common.fake.activities,
+ fr.kazejiyu.ekumi.tests.common.mock
