+++ tests/fr.kazejiyu.ekumi.tests.common/src/fr/kazejiyu/ekumi/tests/common/fake/activities/BrokenActivity.java
@@ -1,0 +1,21 @@
+package fr.kazejiyu.ekumi.tests.common.fake.activities;
+
+import fr.kazejiyu.ekumi.core.ekumi.Activity;
+import fr.kazejiyu.ekumi.core.ekumi.Context;
+import fr.kazejiyu.ekumi.core.ekumi.Status;
+import fr.kazejiyu.ekumi.core.ekumi.impl.ActivityImpl;
+import fr.kazejiyu.ekumi.core.exceptions.EKumiRuntimeException;
+
+/**
+ * An {@link Activity} that sets its status to {@link Status#FAILED}
+ * and throws and Exception when executed. 
+ */
+public class BrokenActivity extends ActivityImpl {
+	
+	@Override
+	public void run(Context context) {
+		setStatus(Status.FAILED);
+		throw new EKumiRuntimeException("I won't run!");
+	}
+	
+}
