+++ tests/fr.kazejiyu.ekumi.tests.common/src/fr/kazejiyu/ekumi/tests/common/fake/activities/DoNothing.java
@@ -1,0 +1,13 @@
+package fr.kazejiyu.ekumi.tests.common.fake.activities;
+
+import fr.kazejiyu.ekumi.core.ekumi.Context;
+import fr.kazejiyu.ekumi.core.ekumi.impl.ActivityImpl;
+
+public class DoNothing extends ActivityImpl {
+	
+	@Override
+	public void run(Context context) {
+		// do nothing
+	}
+
+}
