+++ tests/fr.kazejiyu.ekumi.tests.common/src/fr/kazejiyu/ekumi/tests/common/fake/activities/FakeRunner.java
@@ -1,0 +1,30 @@
+package fr.kazejiyu.ekumi.tests.common.fake.activities;
+
+import fr.kazejiyu.ekumi.core.ekumi.Context;
+import fr.kazejiyu.ekumi.core.ekumi.impl.RunnerImpl;
+
+public class FakeRunner extends RunnerImpl {
+	
+	private int callsToRun = 0;
+	
+	private Context contextOnRun;
+	
+	@Override
+	public void run(Context context) {
+		++callsToRun;
+		contextOnRun = context;
+	}
+	
+	public boolean hasBeenRun() {
+		return callsToRun > 0;
+	}
+	
+	public boolean hasBeenRunOnce() {
+		return callsToRun == 1;
+	}
+	
+	public Context getContextOnRun() {
+		return contextOnRun;
+	}
+
+}
