+++ tests/fr.kazejiyu.ekumi.tests.common/src/fr/kazejiyu/ekumi/tests/common/fake/activities/HopelessActivity.java
@@ -1,0 +1,18 @@
+package fr.kazejiyu.ekumi.tests.common.fake.activities;
+
+import fr.kazejiyu.ekumi.core.ekumi.Activity;
+import fr.kazejiyu.ekumi.core.ekumi.Context;
+import fr.kazejiyu.ekumi.core.ekumi.Status;
+import fr.kazejiyu.ekumi.core.ekumi.impl.ActivityImpl;
+
+/**
+ * An {@link Activity} that sets its status to {@link Status#FAILED} when executed. 
+ */
+public class HopelessActivity extends ActivityImpl {
+	
+	@Override
+	public void run(Context context) {
+		setStatus(Status.FAILED);
+	}
+	
+}
