+++ tests/fr.kazejiyu.ekumi.tests.common/src/fr/kazejiyu/ekumi/tests/common/fake/activities/LoopUntilCancelled.java
@@ -1,0 +1,28 @@
+package fr.kazejiyu.ekumi.tests.common.fake.activities;
+
+import fr.kazejiyu.ekumi.core.ekumi.Context;
+import fr.kazejiyu.ekumi.core.ekumi.Status;
+import fr.kazejiyu.ekumi.core.ekumi.impl.ActivityImpl;
+
+/**
+ * An Activity that increments the value of a specific variable. 
+ */
+public class LoopUntilCancelled extends ActivityImpl {
+	
+	@Override
+	public void run(Context context) {
+		long start = System.currentTimeMillis();
+		
+		while (! context.execution().isCancelled()) {
+			
+			// Prevents an infinite loop when running headless tests
+			if (System.currentTimeMillis() - start > 2000) {
+				setStatus(Status.FAILED);
+				return;
+			}
+		}
+		
+		setStatus(Status.CANCELLED);
+	}
+	
+}
