+++ tests/fr.kazejiyu.ekumi.tests.common/src/fr/kazejiyu/ekumi/tests/common/fake/activities/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Fake activities to ease testing.
+ * 
+ * @author Emmanuel CHEBBI 
+ */
+package fr.kazejiyu.ekumi.tests.common.fake.activities;
